Fixed dependency inversion of router phase #1
