+++ Geography/Geography.csproj
@@ -121,11 +121,13 @@
     <Compile Include="Model\DataModel.cs" />
     <Compile Include="Model\DataModel.Runtime.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Router\GeoRouter.cs" />
-    <Compile Include="Router\GeoRouter.Model.cs" />
-    <Compile Include="Router\GeoRouter.Routing.cs" />
     <Compile Include="UpdateResult.cs" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <ProjectReference Include="..\GeographyRouter\GeographyRouter.csproj">
+      <Project>{cbe819a9-b13b-4c6f-940e-8f3167fa818a}</Project>
+      <Name>GeographyRouter</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
