+++ GeographyRouter/Model.Dependency/IGeoRepository.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace GeographyRouter
+{
+    public interface IGeoRepository
+    {
+        void ResetRouting();
+        List<ILayerElement> GetRoutingSources();
+        void RoutingHitTest(double latitude, double longitude, ref List<ILayerElement> result, bool justNotRoute);
+    }
+}
