+++ GeographyRouter/Model.Dependency/LayerGeographyType.cs
@@ -1,0 +1,10 @@
+﻿namespace GeographyRouter
+{
+    public enum LayerGeographyType
+    {
+        Point = 1,
+        Polyline = 2,
+        Polygon = 3,
+    }
+
+}
